+++ src/util/name.h
@@ -18,6 +18,12 @@ Author: Leonardo de Moura
 
 namespace lean {
 constexpr char const * lean_name_separator = ".";
+constexpr char16_t id_begin_escape = u'«';
+constexpr char16_t id_end_escape = u'»';
+
+bool is_id_first(char const * begin, char const * end);
+bool is_id_rest(char const * begin, char const * end);
+
 enum class name_kind { ANONYMOUS, STRING, NUMERAL };
 /**
    \brief Hierarchical names.
@@ -36,8 +42,8 @@ public:
         };
         void dealloc();
         imp(bool s, imp * p):m_rc(1), m_is_string(s), m_hash(0), m_prefix(p) { if (p) p->inc_ref(); }
-        static void display_core(std::ostream & out, imp * p, char const * sep);
-        static void display(std::ostream & out, imp * p, char const * sep = lean_name_separator);
+        static void display_core(std::ostream & out, imp * p, bool escape, char const * sep);
+        static void display(std::ostream & out, imp * p, bool escape, char const * sep = lean_name_separator);
         friend void copy_limbs(imp * p, buffer<name::imp *> & limbs);
     };
 private:
@@ -124,6 +130,7 @@ public:
     name get_prefix() const { return is_atomic() ? name() : name(m_ptr->m_prefix); }
     /** \brief Convert this hierarchical name into a string. */
     std::string to_string(char const * sep = lean_name_separator) const;
+    std::string escape(char const * sep = lean_name_separator) const;
     /** \brief Size of the this name (in characters). */
     size_t size() const;
     /** \brief Size of the this name in unicode. */
